+++ wolfdesk/modules/base/src/main/kotlin/wolfdesk/base/security/MemberPrincipal.kt
@@ -1,0 +1,5 @@
+package wolfdesk.base.security
+
+data class MemberPrincipal(
+    val memberId: Long,
+)
